chore: reorganize lending ui file structure
