+++ src/components/ui/lending/AssetDetails/SupplyInfoTab.tsx
@@ -1,5 +1,5 @@
 import { UnifiedMarketData } from "@/types/aave";
-import { InfoRow } from "@/components/ui/lending/assetDetails/InfoRow";
+import { InfoRow } from "@/components/ui/lending/AssetDetails/InfoRow";
 import HorizontalProgressBar from "@/components/ui/HorizontalProgressBar";
 import {
   formatBalance,
