+++ src/components/ui/lending/UserContent/UserSupplyContent.tsx
@@ -1,10 +1,11 @@
 "use client";
+
 import React, { useState } from "react";
-import UserBorrowCard from "@/components/ui/lending/UserBorrowCard";
+import UserSupplyCard from "@/components/ui/lending/UserContent/UserSupplyCard";
 import CardsList from "@/components/ui/CardsList";
 import {
-  UserBorrowData,
-  UserBorrowPosition,
+  UserSupplyData,
+  UserSupplyPosition,
   Market,
   UnifiedMarketData,
 } from "@/types/aave";
@@ -12,34 +13,34 @@ import { unifyMarkets } from "@/utils/lending/unifyMarkets";
 import { TokenTransferState } from "@/types/web3";
 import { LendingFilters, LendingSortConfig } from "@/types/lending";
 
-interface UserBorrowContentProps {
-  marketBorrowData: Record<string, UserBorrowData>;
+interface UserSupplyContentProps {
+  marketSupplyData: Record<string, UserSupplyData>;
   activeMarkets: Market[];
-  showZeroBalance?: boolean;
   tokenTransferState: TokenTransferState;
   filters?: LendingFilters;
   sortConfig?: LendingSortConfig | null;
   onSupply: (market: UnifiedMarketData) => void;
   onBorrow: (market: UnifiedMarketData) => void;
-  onRepay: (market: UnifiedMarketData, max: boolean) => void;
+  onWithdraw: (market: UnifiedMarketData, max: boolean) => void;
+  onCollateralToggle: (market: UnifiedMarketData) => void;
 }
 
-interface EnhancedUserBorrowPosition extends UserBorrowPosition {
+interface EnhancedUserSupplyPosition extends UserSupplyPosition {
   unifiedMarket: UnifiedMarketData;
 }
 
 const ITEMS_PER_PAGE = 10;
 
-const UserBorrowContent: React.FC<UserBorrowContentProps> = ({
-  marketBorrowData,
+const UserSupplyContent: React.FC<UserSupplyContentProps> = ({
+  marketSupplyData,
   activeMarkets,
-  showZeroBalance = false,
   tokenTransferState,
   filters,
   sortConfig,
   onSupply,
   onBorrow,
-  onRepay,
+  onWithdraw,
+  onCollateralToggle,
 }) => {
   const [currentPage, setCurrentPage] = useState(1);
 
@@ -52,25 +53,20 @@ const UserBorrowContent: React.FC<UserBorrowContentProps> = ({
     unifiedMarketMap.set(key, market);
   });
 
-  const enhancedPositions: EnhancedUserBorrowPosition[] = [];
+  const enhancedPositions: EnhancedUserSupplyPosition[] = [];
 
-  Object.values(marketBorrowData).forEach((marketData) => {
-    if (marketData.borrows && marketData.borrows.length > 0) {
-      marketData.borrows.forEach((borrow) => {
-        const currencyKey = `${borrow.currency.address.toLowerCase()}-${marketData.chainId}`;
+  Object.values(marketSupplyData).forEach((marketData) => {
+    if (marketData.supplies && marketData.supplies.length > 0) {
+      marketData.supplies.forEach((supply) => {
+        const currencyKey = `${supply.currency.address.toLowerCase()}-${marketData.chainId}`;
         const unifiedMarket = unifiedMarketMap.get(currencyKey);
 
         if (unifiedMarket) {
-          const debtAmount = parseFloat(borrow.debt.usd) || 0;
-          if (!showZeroBalance && debtAmount === 0) {
-            return;
-          }
-
           enhancedPositions.push({
             marketAddress: marketData.marketAddress,
             marketName: marketData.marketName,
             chainId: marketData.chainId,
-            borrow,
+            supply,
             unifiedMarket,
           });
         }
@@ -89,7 +85,7 @@ const UserBorrowContent: React.FC<UserBorrowContentProps> = ({
           .toLowerCase()
           .includes(filterLower) ||
         // Filter by ticker (currency.symbol)
-        position.borrow.currency.symbol.toLowerCase().includes(filterLower) ||
+        position.supply.currency.symbol.toLowerCase().includes(filterLower) ||
         // Also include market name for broader matching
         position.marketName.toLowerCase().includes(filterLower)
       );
@@ -97,51 +93,51 @@ const UserBorrowContent: React.FC<UserBorrowContentProps> = ({
   }
 
   // Apply sorting
-  let userBorrowPositions: EnhancedUserBorrowPosition[];
+  let userSupplyPositions: EnhancedUserSupplyPosition[];
   if (sortConfig) {
-    userBorrowPositions = [...filteredPositions].sort((a, b) => {
+    userSupplyPositions = [...filteredPositions].sort((a, b) => {
       let aValue: number;
       let bValue: number;
 
       switch (sortConfig.column) {
-        case "borrowApy":
-          aValue = parseFloat(a.borrow.apy.value) || 0;
-          bValue = parseFloat(b.borrow.apy.value) || 0;
+        case "supplyApy":
+          aValue = parseFloat(a.supply.apy.value) || 0;
+          bValue = parseFloat(b.supply.apy.value) || 0;
           break;
-        case "userBorrowedValue":
-          aValue = parseFloat(a.borrow.debt.usd) || 0;
-          bValue = parseFloat(b.borrow.debt.usd) || 0;
+        case "userSuppliedValue":
+          aValue = parseFloat(a.supply.balance.usd) || 0;
+          bValue = parseFloat(b.supply.balance.usd) || 0;
           break;
         default:
-          // Default sort by borrowed value (highest first)
-          aValue = parseFloat(a.borrow.debt.usd) || 0;
-          bValue = parseFloat(b.borrow.debt.usd) || 0;
+          // Default sort by supplied value (highest first)
+          aValue = parseFloat(a.supply.balance.usd) || 0;
+          bValue = parseFloat(b.supply.balance.usd) || 0;
           break;
       }
 
       return sortConfig.direction === "asc" ? aValue - bValue : bValue - aValue;
     });
   } else {
-    // Default sort by debt amount (highest first)
-    userBorrowPositions = filteredPositions.sort((a, b) => {
-      const balanceA = parseFloat(a.borrow.debt.usd) || 0;
-      const balanceB = parseFloat(b.borrow.debt.usd) || 0;
+    // Default sort by balance (highest first)
+    userSupplyPositions = filteredPositions.sort((a, b) => {
+      const balanceA = parseFloat(a.supply.balance.usd) || 0;
+      const balanceB = parseFloat(b.supply.balance.usd) || 0;
       return balanceB - balanceA;
     });
   }
 
-  if (userBorrowPositions.length === 0) {
+  if (userSupplyPositions.length === 0) {
     return (
       <div className="text-center py-16">
         <div className="text-[#A1A1AA] text-sm">
-          your borrow positions will be displayed here
+          your supply positions will be displayed here
         </div>
       </div>
     );
   }
 
-  const totalPages = Math.ceil(userBorrowPositions.length / ITEMS_PER_PAGE);
-  const paginatedPositions = userBorrowPositions.slice(
+  const totalPages = Math.ceil(userSupplyPositions.length / ITEMS_PER_PAGE);
+  const paginatedPositions = userSupplyPositions.slice(
     (currentPage - 1) * ITEMS_PER_PAGE,
     currentPage * ITEMS_PER_PAGE,
   );
@@ -154,13 +150,14 @@ const UserBorrowContent: React.FC<UserBorrowContentProps> = ({
     <CardsList
       data={paginatedPositions}
       renderCard={(position) => (
-        <UserBorrowCard
-          key={`${position.marketAddress}-${position.borrow.currency.symbol}`}
+        <UserSupplyCard
+          key={`${position.marketAddress}-${position.supply.currency.symbol}`}
           position={position}
           unifiedMarket={position.unifiedMarket}
           onSupply={onSupply}
           onBorrow={onBorrow}
-          onRepay={onRepay}
+          onWithdraw={onWithdraw}
+          onCollateralToggle={onCollateralToggle}
           tokenTransferState={tokenTransferState}
         />
       )}
@@ -168,9 +165,9 @@ const UserBorrowContent: React.FC<UserBorrowContentProps> = ({
       totalPages={totalPages}
       onPageChange={handlePageChange}
       itemsPerPage={ITEMS_PER_PAGE}
-      totalItems={userBorrowPositions.length}
+      totalItems={userSupplyPositions.length}
     />
   );
 };
 
-export default UserBorrowContent;
+export default UserSupplyContent;
